Remove global.json, so that we no longer pin the dotnet SDK for development-time concerns:
This pinning capability is meant to be used only as-needed, and avoiding it is encouraged so that developers can leverage the latest development tools available. The SDK is meant to be backward compatible with older target frameworks. Note that the solution's TargetFramework selections do still strictly dictate versions pertaining to execution-time concerns.